Require Aloha to be enabled in the User Settings
Aloha is great for editing in the Frontend, but it can also be a problem
e.g. because you cannot follow links in content. This is why it is useful
to let the user decide if he wants to enable Frontend Editing or not.

This change adds a user setting "Enable Frontend Editing". As with every
user setting, it can be overridden by the integrator and therefore
enabling can be forced.