Disallow editing if user only has permissions to show page
In the current implementation, editing rights for content was granted
if the user permission for the page was odd (e.g. 31=admin, 17=
show and edit content). Permission "1" is "Show page" and because if
odd editing of contents was allowed even if the user only had
this permission. This is fixed with an additional check.